credentials from env var in preparation for travis

+++ tests/GitDataRepoTest.php
@@ -5,38 +5,53 @@ namespace GitDataRepo;
 class GitDataRepoTest extends \PHPUnit_Framework_TestCase {
 
     public function setUp() {
-      $fn = "/home/shadi/.mygithubauthjson";
-      $creds = json_decode(file_get_contents($fn),true);
-      # $creds = $creds["http-basic"]["bitbucket.org"];
-      assert(
-        array_key_exists("username",$creds) &&
-        array_key_exists("password",$creds));
-      $remote = "https://".$creds["username"].":".$creds["password"]."@github.com/shadiakiki1986/git-data-repo-testDataRepo";
+      $username = getenv("GITDATAREPO_USERNAME");
+      if(!$username) throw new exception("Please define env var GITDATAREPO_USERNAME");
+      $password = getenv("GITDATAREPO_PASSWORD");
+      if(!$password) throw new exception("Please define env var GITDATAREPO_PASSWORD");
+
+      $remote = GitDataRepo::injectRemoteCredentials(
+        $url,
+        $username,
+        $password);
 
       $repo1 = tempnam("/tmp","test");
       unlink($repo1);
       $gr1 = \Coyl\Git\GitRepo::create($repo1,$remote);
       $this->gdr1 = new GitDataRepo($gr1,$remote,\Monolog\Logger::DEBUG);
 
-      $repo2 = "/home/shadi/Development/git-data-repo-testDataRepo";
-      $gr2 = new \Coyl\Git\GitRepo($repo2,false,false);
+      $repo2 = tempnam("/tmp","test");
+      unlink($repo2);
+      $gr2 = \Coyl\Git\GitRepo::create($repo2,$remote);
       $this->gdr2 = new GitDataRepo($gr2,$remote,\Monolog\Logger::DEBUG);
+
+      // $repo3 = "/home/shadi/Development/git-data-repo-testDataRepo";
     }
 
     public function testGet() {
-#      $bla = $this->gdr2->get("bla");
-#      $this->assertTrue(is_null($bla));
+      $this->gdr1->rm("bla");
 
-      $this->gdr2->set("bla","foo");
-      $bla2 = $this->gdr2->get("bla");
-      $this->assertEquals($bla2,"foo");
+      $bla = $this->gdr1->get("bla");
+      $this->assertTrue(is_null($bla));
 
-#$this->gdr2->push();
-
+      $this->gdr1->set("bla","foo");
       $bla1 = $this->gdr1->get("bla");
       $this->assertEquals($bla1,"foo");
 
+#$this->gdr1->push();
+
+      $bla2 = $this->gdr2->get("bla");
+      $this->assertEquals($bla2,"foo");
+
     }
 
+    public function testInject() {
+      $url = "https://github.com/shadiakiki1986/git-data-repo-testDataRepo";
+      $username = "bla";
+      $password = "bli";
+      $expected = "https://bla:bli@github.com/shadiakiki1986/git-data-repo-testDataRepo";
+      $actual = GitDataRepo::injectRemoteCredentials($url,$username,$password);
+      $this->assertEquals($actual,$expected);
+    }
 }
